+++ src/app/outbound/mod.rs
@@ -1,5 +1,4 @@
 use super::{classify, config::Config, dst::DstAddr, identity, DispatchDeadline};
-use crate::control::destination::{Metadata, Unresolvable};
 use crate::core::listen::ServeConnection;
 use crate::core::resolve::{Resolution, Resolve};
 use crate::proxy::http::{
@@ -7,6 +6,7 @@ use crate::proxy::http::{
     normalize_uri, profiles, retry, router, settings, strip_header,
 };
 use crate::proxy::{self, accept, reconnect, resolve, Server};
+use crate::resolve::{Metadata, Unresolvable};
 use crate::transport::Connection;
 use crate::transport::{self, connect, keepalive, tls};
 use crate::{svc, Addr, NameAddr};
