+++ src/lib.rs
@@ -8,10 +8,10 @@ use linkerd2_metrics as metrics;
 use linkerd2_never::Never;
 use linkerd2_proxy_api as api;
 use linkerd2_proxy_core::{self as core, drain, Error};
+use linkerd2_proxy_resolve as resolve;
 use linkerd2_task as task;
 
 pub mod app;
-pub mod control;
 mod dns;
 pub mod logging;
 mod proxy;
